start ontology update thread on context initialization.
